+++ ocaml-lsp-server/src/code_actions/action_mark_remove_unused.mli
@@ -1,0 +1,3 @@
+val mark : Code_action.t
+
+val remove : Code_action.t