+++ ocaml-lsp-server/src/code_actions/diagnostic_util.ml
@@ -1,0 +1,5 @@
+open Import
+
+let is_unused_var_warning (d : Diagnostic.t) =
+  String.is_prefix d.message ~prefix:"Error (warning 26)"
+  || String.is_prefix d.message ~prefix:"Error (warning 27)"