+++ ocaml-lsp-server/src/ocaml_lsp_server.ml
@@ -311,6 +311,8 @@ let code_action (state : State.t) (params : CodeActionParams.t) =
       ; Action_refactor_open.unqualify
       ; Action_refactor_open.qualify
       ; Action_add_rec.t
+      ; Action_mark_remove_unused.mark
+      ; Action_mark_remove_unused.remove
       ]
   in
   let code_action_results = List.filter_opt code_action_results in